# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/trips": {
      "post": {
        "operationId": "Trips_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      },
      "get": {
        "operationId": "Trips_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive code/name search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Exact resource status filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDataListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{tripId}": {
      "get": {
        "operationId": "Trips_get",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{tripId}/staff": {
      "post": {
        "operationId": "Trips_addStaff",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTripStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{tripId}/stops": {
      "post": {
        "operationId": "Trips_addStop",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTripStopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{tripId}/stops/{tripStopId}/order": {
      "patch": {
        "operationId": "Trips_reorderStop",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tripStopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTripStopDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripStopOrderDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{tripId}/dispatch": {
      "post": {
        "operationId": "Trips_dispatch",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{tripId}/start": {
      "post": {
        "operationId": "Trips_start",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "trips"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMS Sprint 0 API",
    "description": "Frozen Sprint 0 operational API contract. Tenant context is derived from the bearer token.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddTripStaffDto": {
        "type": "object",
        "properties": {
          "staffId": {
            "type": "number"
          },
          "tripRole": {
            "type": "string",
            "enum": [
              "DRIVER",
              "DELIVERY_STAFF",
              "HELPER",
              "RELIEVER"
            ]
          },
          "joinedAt": {
            "type": "string"
          },
          "leftAt": {
            "type": "string"
          }
        },
        "required": [
          "staffId",
          "tripRole"
        ]
      },
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "tripNumber": {
            "type": "string",
            "description": "Generated by the server when omitted"
          },
          "routeId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "tripDate": {
            "type": "string",
            "example": "2026-09-01"
          },
          "shiftCode": {
            "type": "string"
          },
          "vehicleId": {
            "type": "number"
          },
          "primaryStaffId": {
            "type": "number"
          },
          "plannedStartAt": {
            "type": "string"
          },
          "staff": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddTripStaffDto"
            }
          }
        },
        "required": [
          "tripDate"
        ]
      },
      "AddTripStopProductDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number"
          },
          "partyProductId": {
            "type": "number"
          },
          "quantityMode": {
            "type": "string",
            "enum": [
              "FIXED_PLANNED",
              "RETURN_MATCHED",
              "AD_HOC"
            ]
          },
          "plannedQty": {
            "type": "number"
          },
          "forecastQty": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "productId",
          "quantityMode"
        ]
      },
      "AddTripStopDto": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "enum": [
              "ROUTE",
              "ADMIN_ADDED",
              "STAFF_ADDED",
              "CUSTOMER_ORDER",
              "AD_HOC_NEW_CUSTOMER",
              "RESCHEDULED"
            ]
          },
          "sortableOrder": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddTripStopProductDto"
            }
          }
        },
        "required": [
          "partyId",
          "source",
          "sortableOrder",
          "products"
        ]
      },
      "ReorderTripStopDto": {
        "type": "object",
        "properties": {
          "sortableOrder": {
            "type": "number",
            "example": 25
          }
        },
        "required": [
          "sortableOrder"
        ]
      },
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TripData": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "1"
          },
          "tripNumber": {
            "type": "string"
          },
          "routeId": {
            "type": "string",
            "nullable": true
          },
          "tripDate": {
            "type": "string",
            "format": "date"
          },
          "shiftCode": {
            "type": "string",
            "nullable": true
          },
          "vehicleId": {
            "type": "string",
            "nullable": true
          },
          "primaryStaffId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "plannedStartAt": {
            "type": "string",
            "nullable": true
          },
          "actualStartAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "tripId",
          "tripNumber",
          "routeId",
          "tripDate",
          "shiftCode",
          "vehicleId",
          "primaryStaffId",
          "status",
          "plannedStartAt",
          "actualStartAt",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": true
      },
...
      "TripDetailData": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "1"
          },
          "tripNumber": {
            "type": "string"
          },
          "routeId": {
            "type": "string",
            "nullable": true
          },
          "tripDate": {
            "type": "string",
            "format": "date"
          },
          "shiftCode": {
            "type": "string",
            "nullable": true
          },
          "vehicleId": {
            "type": "string",
            "nullable": true
          },
          "primaryStaffId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "plannedStartAt": {
            "type": "string",
            "nullable": true
          },
          "actualStartAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "route": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TripRouteSummaryData"
              }
            ],
            "nullable": true
          },
          "vehicle": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TripVehicleSummaryData"
              }
            ],
            "nullable": true
          },
          "primaryStaff": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TripPrimaryStaffSummaryData"
              }
            ],
            "nullable": true
          },
          "staff": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripStaffAssignmentData"
            }
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripStopData"
            }
          }
        },
        "required": [
          "tripId",
          "tripNumber",
          "routeId",
          "tripDate",
          "shiftCode",
          "vehicleId",
          "primaryStaffId",
          "status",
          "plannedStartAt",
          "actualStartAt",
          "createdAt",
          "updatedAt",
          "route",
          "vehicle",
          "primaryStaff",
          "staff",
          "stops"
        ],
        "additionalProperties": true
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "additionalProperties": true
      },
...
      "TripDataListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripData"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": true
      },
      "TripDetailDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripDetailData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "TripStopOrderDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripStopOrderData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "TripDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
...
    }
  },
...
}